+++ assets/Delete-59b962da.js
@@ -1,1 +1,0 @@
-import{c as t,t as e}from"./index-10f92eee.js";const l=e('<svg><path d="M0 0h24v24H0z" fill="none"></svg>',!1,!0),s=e('<svg><path d="M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zM19 4h-3.5l-1-1h-5l-1 1H5v2h14V4z"></svg>',!1,!0),c=t(()=>[l(),s()],"Delete");export{c as D};